Move system info library to another location
