+++ src/lib/utils/system-info.js
